+++ core/design-system/src/main/java/com/goms/design_system/component/clickable/gomsClickable.kt
@@ -1,4 +1,4 @@
-package com.goms.design_system.component.modifier
+package com.goms.design_system.component.clickable
 
 import androidx.compose.foundation.LocalIndication
 import androidx.compose.foundation.clickable
