+++ core/design-system/src/main/java/com/goms/design_system/component/timer/CountdownTimer.kt
@@ -15,7 +15,7 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.goms.design_system.component.modifier.gomsClickable
+import com.goms.design_system.component.clickable.gomsClickable
 import com.goms.design_system.theme.GomsTheme
 import kotlinx.coroutines.delay
 import java.util.concurrent.TimeUnit
